moved ProvenHeadersConsensusManagerBehavior.cs to Stratis.Bitcoin.Features.Consensus project
